correccion de mensaje de pago
